alawein-technologies-llc/talai/atlas-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-orchestrator"
version = "0.1.0"
description = "Multi-model AI orchestration with intelligent routing and cost optimization"
readme = "README.md"
authors = [
    {name = "Project ATLAS", email = "atlas@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "orchestration", "multi-model", "cost-optimization", "claude", "gpt4", "gemini"]
requires-python = ">=3.9"

dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tiktoken>=0.5.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
atlas = "atlas_orchestrator.cli:app"

[project.urls]
Homepage = "https://github.com/yourorg/atlas-orchestrator"
Documentation = "https://yourorg.github.io/atlas-orchestrator"
Repository = "https://github.com/yourorg/atlas-orchestrator"
"Bug Tracker" = "https://github.com/yourorg/atlas-orchestrator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
